requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.1.0
bleach==3.2.1
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.1
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
docker-pycreds==0.4.0
docopt==0.6.2
emoji==1.2.0
entrypoints==0.3
filelock==3.0.12
fsspec==0.8.4
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
idna==2.8
importlib-metadata==3.1.0
ipykernel==5.3.4
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.17.0
json5==0.9.5
jsonpickle==1.4.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
multidict==5.1.0
munch==2.5.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nlp==0.4.0
nltk==3.5
notebook==6.1.5
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
packaging==20.7
pandas==1.1.4
pandocfilters==1.4.3
parso==0.6.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.4
protobuf==3.14.0
psutil==5.7.0
ptyprocess==0.6.0
py-cpuinfo==7.0.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
Pygments==2.6.1
pymongo==3.11.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-lightning==1.1.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==20.0.0
regex==2020.11.13
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.15.87
sacred==0.8.1
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
Send2Trash==1.5.0
sentencepiece==0.1.94
sentry-sdk==0.19.4
shortuuid==1.0.1
six==1.14.0
sklearn==0.0
smmap==3.0.4
soupsieve==2.0
subprocess32==3.5.4
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.9.4
torch==1.6.0
torchvision==0.7.0
tornado==6.1
tqdm==4.42.1
traitlets==4.3.3
transformers==4.0.0rc1
typing-extensions==3.7.4.3
urllib3==1.25.8
wandb==0.10.11
watchdog==0.10.4
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xxhash==2.0.0
yarl==1.6.3
zipp==3.4.0
